requirements.txt
pandas==1.1.3
Flask==1.1.2
scipy==1.5.2
tensorflow==2.6.0
numpy==1.19.5
matplotlib==3.3.2
connexion==2.7.0
requests==2.24.0
beautifulsoup4==4.9.3
Pillow==8.3.1
